+++ cmd/vegawallet/commands/service_run.go
@@ -129,6 +129,11 @@ func BuildCmdRunService(w io.Writer, handler RunServiceHandler, rf *RootFlags) *
 		"",
 		"Path to the file containing the tokens database passphrase",
 	)
+	cmd.PersistentFlags().Int64Var(&f.TTL,
+		"transaction-ttl",
+		0,
+		"Specify number of blocks during which transactions are considered current",
+	)
 
 	autoCompleteNetwork(cmd, rf.Home)
 
@@ -142,6 +147,7 @@ type RunServiceFlags struct {
 	TokensPassphraseFile   string
 	NoVersionCheck         bool
 	tokensPassphrase       string
+	TTL                    int64
 }
 
 func (f *RunServiceFlags) Validate(rf *RootFlags) error {
@@ -279,7 +285,7 @@ func RunService(w io.Writer, rf *RootFlags, f *RunServiceFlags) error {
 
 	jobRunner := vgjob.NewRunner(context.Background())
 
-	svcURL, errChan, err := serviceStarter.Start(jobRunner, f.Network, f.NoVersionCheck)
+	svcURL, errChan, err := serviceStarter.Start(jobRunner, f.Network, f.NoVersionCheck, uint64(f.TTL))
 	if err != nil {
 		cliLog.Error("Failed to start HTTP server", zap.Error(err))
 		jobRunner.StopAllJobs()